Bug 390699 - Table.showItem is broken on Linux - add build check

+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/widgets/ToolBar.java
@@ -476,7 +476,7 @@ public int getRowCount () {
 	return layout (rect.width, rect.height, false) [0];
 }
 
-boolean hasKeyboardFocus(long /*int*/ inId) {
+boolean hasKeyboardFocus(int /*long*/ inId) {
 	return hasFocus();
 }
